Extract withSchemaKeys to separate file
